requirements.txt
camel-ai[all]
yfinance
tavily-python
python-dotenv
finnhub-python
tushare
akshar
json-repair
